+++ bundlewatch.config.js
@@ -14,8 +14,7 @@ const excludedPackages = [
     'packages/services/f-braze-adapter',
     'packages/tools/generator-component',
     'packages/tools/storybook',
-    'packages/services/f-wdio-utils',
-    'packages/components/organisms/f-status-banner' // Excluding as the footer is pushing this package over the 100kb.
+    'packages/services/f-wdio-utils'
 ];
 
 /**