+++ packages/components/atoms/f-card/src/tenants/nb-NO.js
@@ -1,4 +1,0 @@
-export default {
-    locale: 'nb-NO',
-    cardTitle: 'I am a Card Component (NO)'
-};