+++ packages/components/molecules/f-searchbox/src/components/formElements/tests/FormFullAddressSearchSuggestions.test.js
@@ -1,6 +1,7 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
 import FullAddressSuggestions from '../FormFullAddressSearchSuggestions.vue';
+import { VUEX_MODULE_NAMESPACE } from '../../../services/constants';
 
 const localVue = createLocalVue();
 
@@ -43,7 +44,7 @@ const mockActions = {
 
 const createStore = (state = mockState, actions = mockActions) => new Vuex.Store({
     modules: {
-        searchbox: {
+        [VUEX_MODULE_NAMESPACE]: {
             namespaced: true,
             state,
             actions