+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout.component.spec.js
@@ -1,14 +1,22 @@
-import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
+const Checkout = require('../../../test-utils/component-objects/f-checkout.component');
+
+const checkout = new Checkout();
 
 describe('f-checkout component tests', () => {
     before(() => {
-        browser.url('iframe.html?id=components-organisms--checkout-component&knob-Get%20Checkout%20Url=%2Fcheckout-delivery.json&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json&knob-Create%20Guest%20Url=%2Fcreate-guest.json&knob-Auth%20token=a&knob-Login%20Url=%2Flogin&viewMode=story');
-        CheckoutComponent.waitForCheckoutComponent();
+        const checkoutData = {
+            type: 'delivery',
+            isAuthenticated: true,
+            isValid: true
+        };
+
+        checkout.open(checkoutData);
+        checkout.waitForComponent();
     });
 
     it('should display the f-checkout component', () => {
         // Assert
-        expect(CheckoutComponent.isCheckoutComponentDisplayed()).toBe(true);
+        expect(checkout.isComponentDisplayed()).toBe(true);
     });
 
     it.skip('should submit the checkout form', () => {
@@ -23,9 +31,9 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.selectOrderTime('Wednesday 00:30');
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(addressInfo);
+        checkout.selectOrderTime('Wednesday 00:30');
+        checkout.goToPayment();
 
         // Assert
         // Waiting for route here, so we can grab redirect url and show form submits.
@@ -33,7 +41,7 @@ describe('f-checkout component tests', () => {
 
     it('should display the mandatory "mobileNumber" field', () => {
         // Assert
-        expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
+        expect(checkout.isFieldDisplayed('mobileNumber')).toBe(true);
     });
 
     it('should display a "mobileNumber" error message when an unsupported country code is used in the mobile number field', () => {
@@ -43,11 +51,11 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(addressDetails);
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(addressDetails);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed('mobileNumber')).toBe(true);
+        expect(checkout.isFieldErrorDisplayed('mobileNumber')).toBe(true);
     });
 
     it('should not display a "mobileNumber" error message when a number is formatted with a supported country code', () => {
@@ -57,22 +65,22 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(addressDetails);
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(addressDetails);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed('mobileNumber')).toBe(false);
+        expect(checkout.isFieldErrorDisplayed('mobileNumber')).toBe(false);
     });
 
     it('should display times in ascending order, with default text "As soon as possible" showing first', () => {
         // Act
-        CheckoutComponent.selectOrderTime('As soon as possible');
+        checkout.selectOrderTime('As soon as possible');
 
         // Assert
-        expect(CheckoutComponent.isOrderTimeDropdownDisplayed()).toBe(true);
-        expect(CheckoutComponent.getOrderTimeOptionText(0)).toBe('As soon as possible');
-        expect(CheckoutComponent.getOrderTimeOptionText(1)).toBe('Wednesday 00:45');
-        expect(CheckoutComponent.getOrderTimeOptionText(2)).toBe('Wednesday 01:00');
+        expect(checkout.isOrderTimeDropdownDisplayed()).toBe(true);
+        expect(checkout.getOrderTimeOptionText(0)).toBe('As soon as possible');
+        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 00:45');
+        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 01:00');
     });
 
     it('should prevent a user from writing a note of over 200 characters', () => {
@@ -83,10 +91,10 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        CheckoutComponent.inputUserNote(addressInfo);
+        checkout.inputUserNote(addressInfo);
 
         // Assert
-        expect(CheckoutComponent.userNoteMaxCharacterCount()).toEqual('200');
+        expect(checkout.userNoteMaxCharacterCount()).toEqual('200');
     });
 
     it.skip('should enable a user to submit without adding a note', () => {
@@ -101,15 +109,15 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.selectOrderTime('Wednesday 00:30');
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(addressInfo);
+        checkout.selectOrderTime('Wednesday 00:30');
+        checkout.goToPayment();
 
         // Assert
         // Waiting for route here, so we can grab redirect url and show form submits.
     });
 
     it('should display the switch user link', () => {
-        expect(CheckoutComponent.switchUserLinkIsDisplayed()).toBe(true);
+        expect(checkout.switchUserLinkIsDisplayed()).toBe(true);
     });
 });