+++ packages/components/organisms/f-content-cards/src/components/ContentCards.js
@@ -196,7 +196,8 @@ export default {
                     'Anniversary_Card_1',
                     'Post_Order_Card_1',
                     'Home_Promotion_Card_1',
-                    'Home_Promotion_Card_2'
+                    'Home_Promotion_Card_2',
+                    'Stamp_Card_1'
                 ],
                 brands: this.brands,
                 callbacks: {